+++ exonum-java-binding/core/src/test/java/com/exonum/binding/blockchain/BlockTest.java
@@ -16,12 +16,40 @@
 
 package com.exonum.binding.blockchain;
 
+import static java.nio.charset.StandardCharsets.UTF_8;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import com.exonum.binding.blockchain.Block.Builder;
 import com.exonum.binding.common.hash.HashCode;
+import com.exonum.binding.common.hash.HashFunction;
+import com.exonum.binding.common.hash.Hashing;
 import nl.jqno.equalsverifier.EqualsVerifier;
 import nl.jqno.equalsverifier.Warning;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 class BlockTest {
+
+  @Test
+  void isEmpty() {
+    Block emptyBlock = aBlock()
+        .numTransactions(0)
+        .build();
+
+    assertTrue(emptyBlock.isEmpty());
+  }
+
+  @ParameterizedTest
+  @ValueSource(ints = {1, 2, Integer.MAX_VALUE})
+  void nonEmptyBlock(int numTransactions) {
+    Block nonEmptyBlock = aBlock()
+        .numTransactions(numTransactions)
+        .build();
+
+    assertFalse(nonEmptyBlock.isEmpty());
+  }
 
   @Test
   void verifyEquals() {
@@ -32,5 +60,18 @@ class BlockTest {
         .suppress(Warning.STRICT_HASHCODE)
         .withPrefabValues(HashCode.class, HashCode.fromInt(1), HashCode.fromInt(2))
         .verify();
+  }
+
+  private static Builder aBlock() {
+    HashFunction hashFunction = Hashing.sha256();
+    long blockHeight = 1;
+    return Block.builder()
+        .proposerId(0)
+        .height(blockHeight)
+        .numTransactions(0)
+        .blockHash(hashFunction.hashLong(blockHeight))
+        .previousBlockHash(hashFunction.hashLong(blockHeight - 1))
+        .txRootHash(hashFunction.hashString("transactions at" + blockHeight, UTF_8))
+        .stateHash(hashFunction.hashString("state hash at " + blockHeight, UTF_8));
   }
 }