+++ exonum-java-binding/CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `Block#isEmpty()`
+
 ## [0.6.0]- 2019-05-08
 
 **If you are upgrading an existing Java service, consult